+++ src/main/Recorder.ts
@@ -1527,6 +1527,11 @@ export default class Recorder extends EventEmitter {
     this.monitorCaptureSource.update(settings);
     this.monitorCaptureSource.save();
     this.monitorCaptureSource.enabled = true;
+
+    // Rescale now we're hooked in-case the resolutions don't match. Usually
+    // I'd expect us to get a source callback but sometimes it doesn't seem
+    // to happen.
+    this.scaleVideoSourceSize();
   }
 
   /**
@@ -1924,6 +1929,11 @@ export default class Recorder extends EventEmitter {
 
     console.info('[Recorder] Game capture source configured');
     this.clearFindWindowInterval();
+
+    // Rescale now we're hooked in-case the resolutions don't match. Usually
+    // I'd expect us to get a source callback but sometimes it doesn't seem
+    // to happen.
+    this.scaleVideoSourceSize();
   }
 
   /**
@@ -1967,5 +1977,10 @@ export default class Recorder extends EventEmitter {
 
     console.info('[Recorder] Window capture source configured');
     this.clearFindWindowInterval();
+
+    // Rescale now we're hooked in-case the resolutions don't match. Usually
+    // I'd expect us to get a source callback but sometimes it doesn't seem
+    // to happen.
+    this.scaleVideoSourceSize();
   }
 }